feat(kyc): inscription en ligne, capture faciale et controle d'identite
Parcours d'auto-inscription (ouverture de compte-titres) et controles KYC.

Inscription (V22/V23)
- registration_dossiers, dossier_titulaires/_adresses/_contacts, justificatifs,
  face_captures, convention_versions, convention_acceptations.
- RegistrationController : creation de compte + dossier, capture faciale,
  convention. FileStorageService (stockage hors webroot, SHA-256,
  anti path-traversal). AuthSessionService (emission jeton/cookie partagee).
- Compte especes : 22 chiffres, format « 10005 0001 xxxxxxxxxxx xx ».
- La vivacite est RE-CONTROLEE cote serveur : le front ne peut pas forcer une
  capture.

Comparaison faciale a la soumission d'ordre (V24)
- face_captures.descriptor (empreinte 128-d) + order_face_checks.
- FaceMatcher : le SERVEUR seul calcule la distance et applique le seuil (0,6).
  Un client modifie ne peut pas s'auto-declarer « reconnu ».
- OrderFaceCheckController : le controle intervient APRES la soumission et
  repond toujours 200. Un visage different ne refuse jamais l'ordre — il le
  signale au back-office, qui tranche (revue humaine).

Piece d'identite (V25)
- justificatifs + document_type, cote, ocr_texte, nettete, largeur, hauteur,
  verification_status. Le serveur conserve l'image ET le texte lu : le controle
  navigateur filtre la saisie, le back-office reste l'autorite.

Back-office
- AdminRegistrationController : file des captures, image + reperes 468 pts,
  validation/rejet ; acces a la photo depuis le dossier client (by-user).

Documentation OpenAPI
- springdoc + Swagger UI (/swagger-ui.html), schema Bearer JWT.
  Desactivable en production (SPRINGDOC_ENABLED=false).
- CSP assouplie pour la seule UI Swagger ; l'API garde default-src 'none'.

Qualite
- Code smells 130 -> 13 ; couverture 77,8 % -> 84 %.
- Correction d'un test instable (JwtServiceTest) : l'alteration du dernier
  caractere base64url de la signature etait parfois sans effet.

658 tests verts (mvn verify).

+++ src/main/java/cm/afriland/titres/web/PortfolioController.java
@@ -192,7 +192,7 @@ public class PortfolioController {
                 long valNom = getValeurNominale(emId);
                 long montant = row.montantAdjuge() > 0
                         ? row.montantAdjuge()
-                        : (long) volume * valNom;
+                        : volume * valNom;
                 if (montant <= 0) montant = valNom;
 
                 String compteTitres = (row.numeroCompte() != null && !row.numeroCompte().isBlank())
